agri-gyaan-setu/agri-gyaan-setu/backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
joblib
scikit-learn
pandas
numpy
mysqlclient
dj-database-url
requests
